Implemented Authorizetion using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -10,6 +10,10 @@ use App\Models\User;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -70,6 +74,7 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         //
+        $this->authorize('update', $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -83,6 +88,7 @@ class QuestionsController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
         return redirect()->route('question.index')->with('success','Question updated successfully.');
 
@@ -96,7 +102,7 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        //
+        $this->authorize('delete', $question);
         $question->delete();
         return redirect()->route('question.index')->with('success','Question deleted successfully.');
     }
